Refactored clientside code into directories
