+++ client/src/components/AdminPage/AdminPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+class AdminPage extends React.Component {
+    render() {
+
+        <div>
+            Admin page
+        </div>
+    }
+}
