+++ client/src/components/LoginPage/LookUpAccountDialog.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppMode from '../AppMode.js';
+import AppMode from '../../AppMode.js';
 
 class LookUpAccountDialog extends React.Component {
