+++ client/src/components/RoundsPage/RoundForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppMode from './../AppMode.js';
+import AppMode from '../../AppMode.js';
 
 class RoundForm extends React.Component {
   constructor(props) {
