+++ client/src/components/common/SideMenu.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppMode from './../AppMode.js'
+import AppMode from '../../AppMode.js'
 
 class SideMenu extends React.Component {
